# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Zenstats — 自托管网站分析服务，提供隐私友好的 Web 分析 API。\n\n## 认证方式\n- **BearerAuth**: 在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`，用于管理类 API\n- **APIKeyAuth**: 在 Authorization 头中传入 API Key，用于统计查询 API\n\n## 统一响应格式\n成功: `{ \"code\": 200, \"message\": \"success\", \"data\": ... }`\n失败: `{ \"code\": \u003cstatus\u003e, \"message\": \"...\", \"error\": \"...\" }`\n\n## API 分组\n- **认证** — 登录、注册、令牌刷新、密码重置、邮箱验证、系统初始化\n- **站点管理** — 站点 CRUD、域名验证、屏蔽规则（IP/域名/国家）\n- **统计分析** — 聚合指标、时间序列、维度细分、实时访客、CSV 导出\n- **漏斗分析** — 漏斗 CRUD + 转化率分析\n- **目标管理** — 转化目标 CRUD\n- **事件采集** — 前端埋点 SDK 上报端点\n- **数据导入** — GA4 历史数据导入与查询\n- **用户** — 自定义搜索引擎、子账号管理、配额查询、个人资料\n- **API Key** — API Key 创建、列表、删除\n- **管理员** — 用户/站点/套餐管理、系统配置、系统统计\n- **健康检查** — 服务及依赖连通性检查",
        "title": "Zenstats API",
        "termsOfService": "https://github.com/zenstats/zenstats",
        "contact": {
            "name": "Zenstats",
            "url": "https://github.com/zenstats/zenstats",
            "email": "wrpota@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auth/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "已登录用户修改自己的密码，需要提供旧密码。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "旧密码不正确或未登录",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/forgot-password": {
            "post": {
                "description": "输入邮箱发送密码重置链接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "忘记密码",
                "parameters": [
                    {
                        "description": "忘记密码请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "邮件发送成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/reset-password": {
            "post": {
                "description": "通过重置 token 设置新密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "密码重置成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误或 token 无效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/send-verification": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "向当前用户发送邮箱验证邮件",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "发送验证邮件",
                "responses": {
                    "200": {
                        "description": "发送成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/verification-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前用户的邮箱验证状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "获取邮箱验证状态",
                "responses": {
                    "200": {
                        "description": "验证状态",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/types.VerificationStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify-email": {
            "get": {
                "description": "通过验证链接中的 token 验证邮箱",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "验证邮箱",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证 token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证成功",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "token 无效或过期",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.ForgotPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "auth.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "token"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 6
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "response.ErrorResponse": {
            "description": "接口返回的错误响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "error": {
                    "description": "错误详情",
                    "type": "string"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "description": "接口返回的成功响应信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应信息",
                    "type": "string"
                }
            }
        },
...
        "types.ChangePasswordRequest": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 8
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "types.VerificationStatus": {
            "type": "object",
            "properties": {
                "email_verified": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "在 Authorization 头中传入 API Key（无需 Bearer 前缀）。在站点管理页面创建 API Key。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "在 Authorization 头中传入 `Bearer \u003cJWT Token\u003e`。获取方式：调用 /auth/login 或 /auth/register。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}